requirements.txt
pyserial>=3.5
matplotlib>=3.7.0
numpy>=1.24.0
pywin32>=306
PyYAML>=6.0
pyinstaller>=5.0
Pillow>=9.0.0
